removed substr() from getNamedRoute() on route matching
The substr() method was removing the / from the route.  This created a relative link vs an absolute link, so when I was on the home page all the links worked fine.  When I hit a page with a `subfolder` it was relative and the links did not work as expected.

Registered: 
$router->register('home','MAP /', '\Controller\Main');
$router->register('test','MAP /test', '\Controller\Test');

Templates:
<a route="home">Home</a>
<a route="test">Tests</a>

Results for /
<a href="">Home</a>
<a href="test">Tests</a>

Results for /test
<a href="test">Home</a>
<a href="test">Tests</a>